Susceptible-Exposed-Infectious-Removed (SEIR) Model
Source: R/SEIR.R

This function is used by specific disease models in EPIRICE to model disease severity.

SEIR(wth, emergence, onset = 15, duration = 120, rhlim = 90,
rainlim = 5, wetness = 0, initSites, initInfection = 1, ageRc, tmpRc, rhRc,
baseRc, latrans, inftrans, siteMax, AGGR, RRPhysiolSenesc, RRG,
SenescType = 1)

Arguments

wth
a data frame of weather on a daily time-step containing data with the following field names.

YEAR

Year in YYYY format

DOY

Numeric day of year, e.g. 1 - 365

T2M

Mean daily temperature

T2MN

Minimum daily temperature

T2MX

Maximum daily temperature

RH2M

Relative humidity

RAIN

Precipitation

emergence
expected date of plant emergence entered in YYYY/MM/DD format (usually supplied through the disease model function)

onset
expected number of days until the onset of disease after emergence date

duration
simulation duration

rhlim
threshold to decide whether leaves are wet or not (usually 90%)

rainlim
threshold to decide whether leaves are wet or not

wetness
simulate RHmax or rain threshold (0) or leaf wetness duration (1)

init_sites

init_infection

age_rc

tmp_rc

rh_rc

base_rc
corrected basic infection rate

latrans
latent period

inftrans
infectious period

site_max

aggr

rr_physiol_senesc
relative rate of physiological senescence

rrg

senesc_type

Details

SEIR is called by the following specific disease models: predict_leaf_blast, predict_bacterial_blight, predict_brown_spot, predict_sheath_blight