Predict bacterial blight area under the disease progress curve (AUDPC)

A dynamic mechanistic simulation of bacterial blight disease of rice. The model is driven by daily weather data. Adapted from cropsim package version 0.2.0-5 by Adam H. Sparks - USQ CCH. Original model development: Serge Savary & Rene Pangga (IRRI). Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario (IRRI).

bacterial_blight(wth, emergence = "2000-05-15", ...)

Arguments

wth
- Weather data source

emergence
- Expected date of crop emergence

...
Additional arguments - See SEIR

Value

A raster of AUDPC values for bacterial blight based on given weather data

References

Adhikari, T.B., 1991. Effects of rice genotype and environment on bacterial blight progression. PhD thesis, University of the Philippines at Los Baños, 143 p.

Nayak, P., Suriya Rao, A.V., Chakrabarti, N.K., 1987. Components of resistance to bacterial blight disease of rice. Journal of Phytopathology 119:312-318.

Baw A. and Mew, T.W., 1988. Scoring systems for evaluating rice varietal resistance to bacterial blight (BB): lesion size by growth stage. International Rice Research Newsletter 13:10-11.

Luo Wei-Hong, 1996. Simulation and measurement of leaf wetness formation in paddy rice crops. PhD, Wageningen Agricultural University, 87 p.

Horino, O., Mew, T.W., Yamada, T., 1982. The effect of temperature on the development of bacterial leaf blight on rice. Annals of the Phytopathological Society of Japan 48: 72-75

Medalla, E. 1992. Characterization of resistance of IR cultivars to two races of Xanthomonas oryzae pv. oryzae. Unpublished M.S. Thesis, University of the Philippines at Los Baños, 81 p.

See also

leaf_blast, sheath_blight, brown_spot, tungro

Examples

wth <- readr::read_table(system.file("extdata", "daily_weather_28368.nasa",
package = "epiRice"), skip = 13)

#> Error: '' does not exist in current working directory ('/Users/asparks/Development/EPIRICE/docs/reference').

bb <- bacterial_blight(wth, onset = 20, duration = 120)

#> Error in subset(wth@w, wth@w$date >= emergence - 1): object 'wth' not found

plot(bb, type = 2)

#> Error in plot(bb, type = 2): object 'bb' not found