Get weather data from NASA POWER API

Get weather data from NASA POWER API

get_wth(lonlat, dates)

Arguments

lonlat
A numeric vector of geographic coordinates for a cell or region entered as x, y coordinates.

dates
A character vector of start and end dates in that order.

Value

A data.table of weather data, dates and geolocation information (LAT/LON values) suitable for use in epirice with the following columns:

YYYYMMDD:	Date as Year Month Day (ISO8601).

DOY:	Consecutive day of year, commonly called "Julian date".

TM:	Mean daily temperature (°C).

TN:	Minimum daily temperature (°C).

TX:	Maximum daily temperature (°C).

TDEW:	Mean daily dew point temperature (°C).

RH:	Mean daily relative humidity (%).

RAIN:	Mean daily rainfall (mm).

LAT:	Latitude of area of interest.

LON:	Longitude of area of interest.

Details

This function is just a wrapper for the nasapower get_power function with predefined parameters suitable for use in EPIRICE.

Author

Adam H. Sparks

Examples

# \donttest{
# get weather for IRRI Zeigler Experiment Station in wet season 2000
(wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
))

#>        YYYYMMDD DOY    TM    TN    TX  TDEW    RH  RAIN      LAT      LON
#>   1: 2000-06-30 182 25.94 29.19 22.97 23.63 87.22 11.36 14.67741 121.2556
#>   2: 2000-07-01 183 25.34 28.31 23.65 23.76 91.07 24.87 14.67741 121.2556
#>   3: 2000-07-02 184 25.99 29.91 23.28 23.40 85.71 17.63 14.67741 121.2556
#>   4: 2000-07-03 185 25.35 27.23 24.00 24.32 94.01 33.52 14.67741 121.2556
#>   5: 2000-07-04 186 25.58 27.35 24.20 24.42 93.28 16.21 14.67741 121.2556
#>  ---                                                                     
#> 181: 2000-12-27 362 24.27 26.20 22.65 23.06 92.93 26.32 14.67741 121.2556
#> 182: 2000-12-28 363 24.39 26.87 23.16 23.08 92.33  6.67 14.67741 121.2556
#> 183: 2000-12-29 364 24.39 27.46 22.58 22.94 91.57  7.19 14.67741 121.2556
#> 184: 2000-12-30 365 24.96 28.50 22.60 22.67 87.13  2.97 14.67741 121.2556
#> 185: 2000-12-31 366 24.22 28.49 21.26 21.95 87.21  3.04 14.67741 121.2556

# }