Predict brown spot area under the disease progress curve (AUDPC)

A dynamic mechanistic simulation of rice brown spot. The model is driven by daily weather data. Adapted from cropsim package version 0.2.0-5 by Adam H. Sparks - USQ CCH. Original model development: Serge Savary & Rene Pangga (IRRI). Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario (IRRI).

predict_brown_spot(wth, emergence = "2000-05-15", ...)

Arguments

wth
Weather data with a daily time-step, normally NASA-POWER, but any data that has the following fields will work:

YEAR

Year in YYYY format

DOY

Numeric day of year, e.g. 1 - 365

T2M

Mean daily temperature

T2MN

Minimum daily temperature

T2MX

Maximum daily temperature

RH2M

Relative humidity

RAIN

Precipitation

emergence
Expected date of crop emergence

...
Additional arguments, see SEIR

Value

A raster of AUDPC values for brown spot based on given weather data

References

Klomp, A.O., 1977. Early senescence of rice and Dreschslera oryzae in the Wageningen polder, Surinam. PhD thesis, 97p.

Levy, Y. and Cohen, Y., 1980. Sporulation of Helminthosporium turcicum on sweet corn: Effects of temperature and dew period. Canadian Journal of Plant Pathology 2:65-69.

Padmanabhan, S.Y. and Ganguly, D. 1954. Relation between the age of rice plant and its susceptibility to Helminthosporium and blast disease. Proceedings of the Indian Academy of Sciences B 29:44-50.

Sarkar, A.K. and Sen Gupta, P.K., 1977. Effect of temperature and humidity on disease development and sporulation of Helminthosporium oryzae on rice. Indian Phytopathology 30:258-259.

Luo Wei-Hong, 1996. Simulation and measurement of leaf wetness formation in paddy rice crops. PhD, Wageningen Agricultural University, 87 p.

Waggoner. P.E., Horsfall, J.G., and Lukens, R.J. 1972. EPIMAY. A Simulator of Southern Corn Leaf Blight. Bulletin of the Connecticut Experiment Station, New Haven, 85 p.

See also

predict_leaf_blast, predict_bacterial_blight, predict_brown_spot, predict_sheath_blight

Examples

wth <-
bs <- predict_brown_spot(wth, emergence = "2000-05-15")

#> Error in subset(wth@w, wth@w$DOY >= emergence - 1): object 'wth' not found

plot(bs, type = 2)

#> Error in plot(bs, type = 2): object 'bs' not found