Predict tungro area under the disease progress curve (AUDPC)

A dynamic mechanistic simulation of tungro disease of rice. The model is driven by daily weather data. Adapted from cropsim package version 0.2.0-5 by Adam H. Sparks - USQ CCH. Original model development: Serge Savary & Rene Pangga (IRRI). Original R implementation by Robert J. Hijmans, Rene Pangga, & Jorrel Aunario (IRRI).

predict_tungro(wth, emergence = "2000-05-15", ...)

Arguments

wth
Weather data with a daily time-step, normally NASA-POWER, but any data that has the following fields will work:

YEAR

Year in YYYY format

DOY

Numeric day of year, e.g. 1 - 365

T2M

Mean daily temperature

T2MN

Minimum daily temperature

T2MX

Maximum daily temperature

RH2M

Relative humidity

RAIN

Precipitation

emergence
Expected date of crop emergence

...
Additional arguments, see SEIR

Value

A raster of AUDPC values for tungro based on given weather data

References

Tiongco, E.R., Cabunagan, R.C., Flores, Z.M., Hibino, H., and Koganezawa, H., 1993. Serological monitoring of rice tungro disease development in the field: its implication in disease management. Plant Disease 77:877-882.

Rivera, C.T. and Ou, S.H., 1965. Leafhopper transmission of tungro disease of rice. Plant Disease Reporter 49:127-131.

Ling, K.C., Palomar, M.K., 1966. Studies on rice plants infected with the tungro virus at different ages. Philippines Agriculturist 50:165-177.

Ling, K.C., and Tiongco, E.R., Effect of temperature on the transmission of rice tungro virus by Nephotettix virescens. Philippine Phytopathology 11:46-57.

See also

#' predict_leaf_blast, predict_bacterial_blight, predict_brown_spot, predict_sheath_blight

Examples

wth <-
tg <- predict_tungro(wth, emergence = "2000-05-15")

#> Error in subset(wth@w, wth@w$DOY >= emergence - 1): object 'wth' not found

plot(tg, type = 2)

#> Error in plot(tg, type = 2): object 'tg' not found