~~function to do ... ~~

~~ A concise (1-5 lines) description of what the function does. ~~

spatSim(raster, model, starts, verbose=FALSE, ...)
spatSimFlex(raster, pdateraster, model, years, period=14, periodpt=7, skipzero=TRUE, verbose=FALSE, ...)
getLandCells(tablename, odbcname)

Arguments

raster
~~Describe yields here~~

model
~~Describe fact here~~

starts
~~Describe fact here~~

years
~~Describe fact here~~

verbose
~~Describe fact here~~

pdateraster
~~Describe fact here~~

period
~~Describe fact here~~

periodpt
~~Describe fact here~~

skipzero
~~Describe fact here~~

odbcname
~~Describe fact here~~

tablename
~~Describe fact here~~

...
~~Describe fact here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1

Description of 'comp1'

comp2

Description of 'comp2'

...
References

~put references to the literature/web site here ~

Note

~~further notes~~

See also

~~objects to See Also as help, ~~~

Examples