Susceptible-Exposed-Infectious-Removed (SEIR) model

This function is used by specific disease models in EPIRICE to model disease severity of several rice diseases. It should be generic enough to port to other pathosystems.

SEIR(
  wth,
  emergence,
  onset = 15,
  duration = 120,
  rhlim = 90,
  rainlim = 5,
  wetness = 0,
  init_sites,
  init_infection = 1,
  age_rc,
  tmp_rc,
  rh_rc,
  base_rc,
  latrans,
  inftrans,
  site_max,
  aggr,
  rr_physiol_senesc,
  rrg,
  senesc_type = 1
)

Arguments

wth
a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD Date in YYYY-MM-DD format

DOY Numeric day of year, e.g. 1 - 365

T2M Mean daily temperature

T2MN Minimum daily temperature

T2MX Maximum daily temperature

RH2M Relative humidity

RAIN Precipitation

emergence
expected date of plant emergence entered in YYYY-MM-DD format (usually supplied through the disease model function)

onset
expected number of days until the onset of disease after emergence date

duration
simulation duration

rhlim
threshold to decide whether leaves are wet or not (usually 90 pct)

rainlim
threshold to decide whether leaves are wet or not

wetness
simulate RHmax or rain threshold (0) or leaf wetness duration (1)

init_sites
Number of initial infection sites

init_infection
NA

age_rc
NA

tmp_rc
NA

rh_rc
NA

base_rc
corrected basic infection rate

latrans
latent period

inftrans
infectious period

site_max
NA

aggr
NA

rr_physiol_senesc
relative rate of physiological senescence

rrg
NA

senesc_type
NA

Value

A data.table containing the following columns

simday:	Zero indexed day of simulation that was run.

dates:	Date of simulation.

sites:	Total number of sites present on day "x".

latent:	Number of latent sites present on day "x".

infectious:	Number of infectious sites present on day "x".

removed:	Number of removed sites present on day "x".

senesced:	Number of senesced sites present on day "x".

rateinf:	...

rtransfer:	...

rgrowth:	...

rsenesced:	...

diseased:	...

severity:	Disease severity or incidence (for tungro).

Details

SEIR is called by the following specific disease models: predict_leaf_blast, predict_bacterial_blight, predict_brown_spot, predict_sheath_blight

Author

Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario, Adam H. Sparks, Aji Sukarta

Examples

# \donttest{
# get weather for IRRI Zeigler Experiment Station in dry season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-01-15", "2000-05-31")
)

# provide suitable values for brown spot severity
age_rc <-
  cbind(0:6 * 20, c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
tmp_rc <-
  cbind(15 + (0:5) * 5, c(0, 0.06, 1.0, 0.85, 0.16, 0))
rh_rc <- cbind(0:8 * 3,
             c(0, 0.12, 0.20, 0.38, 0.46, 0.60, 0.73, 0.87, 1.0))
emergence <- "2000-01-15"

x <- SEIR(
  wth = wth,
  emergence = emergence,
  age_rc = age_rc,
  tmp_rc = tmp_rc,
  rh_rc = rh_rc,
  base_rc = 0.61,
  latrans = 6,
  inftrans = 19,
  init_sites = 600,
  aggr = 1,
  site_max = 100000,
  rr_physiol_senesc = 0.01,
  rrg = 0.1
)
# }