Susceptible-Exposed-Infectious-Removed (SEIR) model framework

This function is originally used by specific disease models in EPIRICE to model disease severity of several rice diseases. Given proper values it can be used with other pathosystems as well.

SEIR(
  wth,
  emergence,
  onset,
  duration,
  rhlim = 90,
  rainlim = 5,
  wetness_type = 0,
  H0,
  I0 = 1,
  RcA,
  RcT,
  RcW,
  RcOpt,
  p,
  i,
  Sx,
  a,
  RRS,
  RRG
)

Arguments

wth
a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD:	Date as Year Month Day (ISO8601).

DOY:	Consecutive day of year, commonly called "Julian date".

TM:	Mean daily temperature (°C).

TN:	Minimum daily temperature (°C).

TX:	Maximum daily temperature (°C).

TDEW:	Mean daily dew point temperature (°C).

RH:	Mean daily relative humidity (%).

RAIN:	Mean daily rainfall (mm).

emergence
expected date of plant emergence entered in YYYY-MM-DD format. From Table 1 Savary et al. 2012.

onset
expected number of days until the onset of disease after emergence date. From Table 1 Savary et al. 2012.

duration
simulation duration (growing season length). From Table 1 Savary et al. 2012.

rhlim
threshold to decide whether leaves are wet or not (usually 90 %). From Table 1 Savary et al. 2012.

rainlim
threshold to decide whether leaves are wet or not. From Table 1 Savary et al. 2012.

wetness_type
simulate RHmax or rain threshold (0) or leaf wetness duration (1). From Table 1 Savary et al. 2012.

H0
initial number of plant's healthy sites. From Table 1 Savary et al. 2012.

I0
initial number of infective sites. From Table 1 Savary et al. 2012.

RcA
crop age curve for pathogen optimum. From Table 1 Savary et al. 2012.

RcT
temperature curve for pathogen optimum. From Table 1 Savary et al. 2012.

RcW
relative curve for pathogen optimum. From Table 1 Savary et al. 2012.

RcOpt
potential basic infection rate corrected for removals. From Table 1 Savary et al. 2012.

p
duration of latent period. From Table 1 Savary et al. 2012.

i
duration of infectious period. From Table 1 Savary et al. 2012.

Sx
maximum number of sites. From Table 1 Savary et al. 2012.

a
aggregation coefficient. From Table 1 Savary et al. 2012.

RRS
relative rate of physiological senescence. From Table 1 Savary et al. 2012.

RRG
relative rate of growth. From Table 1 Savary et al. 2012.

Value

A data.table containing the following columns

simday:	Zero indexed day of simulation that was run.

dates:	Date of simulation.

sites:	Total number of sites present on day "x".

latent:	Number of latent sites present on day "x".

infectious:	Number of infectious sites present on day "x".

removed:	Number of removed sites present on day "x".

senesced:	Number of senesced sites present on day "x".

rateinf:	Rate of infection.

rtransfer:	Rate of transfer from latent to infectious sites.

rgrowth:	Rate of growth of healthy sites.

rsenesced:	Rate of senescence of healthy sites.

diseased:	Number of diseased (latent + infectious + removed) sites.

severity:	Disease severity or incidence (for tungro).

Details

SEIR is called by the following specific disease modelling functions:

predict_bacterial_blight,

predict_brown_spot,

predict_leaf_blast,

predict_sheath_blight,

predict_tungro

References

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling and mapping potential epidemics of rice diseases globally. Crop Protection, Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI: http://dx.doi.org/10.1016/j.cropro.2011.11.009.>

Author

Serge Savary, Ireneo Pangga, Robert Hijmans, Jorrel Khalil Aunario, Adam H. Sparks, Aji Sukarta

Examples

# \donttest{
# get weather for IRRI Zeigler Experiment Station in wet season 2000
wth <- get_wth(
  lonlat = c(121.25562, 14.6774),
  dates = c("2000-06-30", "2000-12-31")
)

# provide suitable values for brown spot severity
RcA <-
  cbind(0:6 * 20, c(0.35, 0.35, 0.35, 0.47, 0.59, 0.71, 1.0))
RcT <-
  cbind(15 + (0:5) * 5, c(0, 0.06, 1.0, 0.85, 0.16, 0))
RcW <- cbind(0:8 * 3,
             c(0, 0.12, 0.20, 0.38, 0.46, 0.60, 0.73, 0.87, 1.0))
emergence <- "2000-07-15"

(SEIR(
  wth = wth,
  emergence = emergence,
  onset = 20,
  duration = 120,
  RcA = RcA,
  RcT = RcT,
  RcW = RcW,
  RcOpt = 0.61,
  p =  6,
  i = 19,
  H0 = 600,
  a = 1,
  Sx = 100000,
  RRS = 0.01,
  RRG = 0.1
))

#>      simday      dates      sites    latent infectious  removed    senesced
#>   1:      0 2000-07-14   600.0000    0.0000      0.000   0.0000     0.00000
#>   2:      1 2000-07-15   653.6400    0.0000      0.000   0.0000     6.00000
#>   3:      2 2000-07-16   712.0404    0.0000      0.000   0.0000    12.53640
#>   4:      3 2000-07-17   775.6170    0.0000      0.000   0.0000    19.65680
#>   5:      4 2000-07-18   844.8209    0.0000      0.000   0.0000    27.41297
#>  ---                                                                       
#> 117:    116 2000-11-07 88004.1105  420.0799   1001.359 120.1544 52488.07402
#> 118:    117 2000-11-08 88044.0904  234.2597   1187.179 120.1544 53368.11512
#> 119:    118 2000-11-09 87411.2473  903.5810   1187.179 120.1544 54248.55603
#> 120:    119 2000-11-10 86738.0725 1348.7742   1394.690 146.9034 55149.41748
#> 121:    120 2000-11-11 85973.0319 2146.0432   1394.690 146.9034 56016.79821
#>       rateinf rtransfer   rgrowth  rsenesced diseased severity
#>   1:   0.0000    0.0000  59.64000   6.000000    0.000 0.000000
#>   2:   0.0000    0.0000  64.93675   6.536400    0.000 0.000000
#>   3:   0.0000    0.0000  70.69703   7.120404    0.000 0.000000
#>   4:   0.0000    0.0000  76.96012   7.756170    0.000 0.000000
#>   5:   0.0000    0.0000  83.76837   8.448209    0.000 0.000000
#>  ---                                                          
#> 117:   0.0000  185.8202 920.02103 880.041105 1541.593 1.589522
#> 118: 669.3212    0.0000 916.91899 880.440904 1541.593 1.588812
#> 119: 679.4530  234.2597 907.13977 900.861455 2210.915 2.335993
#> 120: 797.2690    0.0000 899.60910 867.380725 2890.368 3.065956
#> 121: 809.6519    0.0000 888.90367 859.730319 3687.637 3.955518

# }