Susceptible-Exposed-Infectious-Removed (SEIR) Model

This function is used by specific disease models in EPIRICE to model disease severity.

SEIR(wth, emergence, onset = 15, duration = 120, rhlim = 90,
rainlim = 5, wetness = 0, initSites, initInfection = 1, ageRc, tmpRc, rhRc,
baseRc, latrans, inftrans, siteMax, AGGR, RRPhysiolSenesc, RRG,
SenescType = 1)

Arguments

wth
- daily weather data frame containing relative humidity (relh), minimum temperature (tmin), and maximum temperature (tmax)

emergence
- expected date of plant emergence

onset
-expected number of days until the onset of disease after emergence date

duration
- simulation duration

rhlim
-threshold to decide whether leaves are wet or not (usually 90%)

rainlim
- threshold to decide whether leaves are wet or not

wetness
-simulate RHmax or rain threshold (0) or leaf wetness duration (1)

init_sites
-

init_infection
-

age_rc
-

tmp_rc
-

rh_rc
-

base_rc
- corrected basic infection rate

latrans
- latent period

inftrans
- infectious period

site_max
-

aggr
-

rr_physiol_senesc
- relative rate of physiological senescence

rrg
-

senesc_type
-

Details

SEIR is called by the following specific disease models: leaf_blast, bacterial_blight, brown_spot, sheath_blight, tungro